refine truncated
