+++ src/common/src/cast/datetime.rs
@@ -1,3 +1,17 @@
+// Copyright 2023 RisingWave Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 use chrono::{NaiveDate, NaiveDateTime, NaiveTime};
 use snafu::{ResultExt, Snafu};
 use speedate::{Date as SpeedDate, DateTime as SpeedDateTime, Time as SpeedTime};
@@ -7,6 +21,7 @@ use crate::types::{Date, Time, Timestamp, Timestamptz};
 
 type Result<T> = std::result::Result<T, DateTimeCastError>;
 
+/// Wrapper for `speedate::ParseError` to implement `Error` trait.
 #[derive(Snafu, Debug)]
 #[snafu(display("{}", inner.get_documentation().unwrap()))]
 pub struct SpeedateParseError {
